+++ Sources/TuistLoader/Models+ManifestMappers/SwiftPackageManagerDependencies+ManifestMapper.swift
@@ -12,7 +12,16 @@ extension TuistGraph.SwiftPackageManagerDependencies {
         manifest: ProjectDescription.SwiftPackageManagerDependencies,
         generatorPaths: GeneratorPaths
     ) throws -> Self {
-        let packages = try manifest.packages.map { try TuistGraph.Package.from(manifest: $0, generatorPaths: generatorPaths) }
+        let packagesOrManifest: TuistGraph.PackagesOrManifest
+        switch manifest.packagesOrManifest {
+        case let .packages(packages):
+            packagesOrManifest = .packages(try packages.map { try TuistGraph.Package.from(
+                manifest: $0,
+                generatorPaths: generatorPaths
+            ) })
+        case .manifest:
+            packagesOrManifest = .manifest
+        }
         let productTypes = manifest.productTypes.mapValues { TuistGraph.Product.from(manifest: $0) }
         let baseSettings = try TuistGraph.Settings.from(manifest: manifest.baseSettings, generatorPaths: generatorPaths)
         let targetSettings = manifest.targetSettings.mapValues { TuistGraph.SettingsDictionary.from(manifest: $0) }
@@ -21,7 +30,7 @@ extension TuistGraph.SwiftPackageManagerDependencies {
             .mapValues { .from(manifest: $0) }
 
         return .init(
-            packages,
+            packagesOrManifest,
             productTypes: productTypes,
             baseSettings: baseSettings,
             targetSettings: targetSettings,