+++ Tests/ProjectDescriptionTests/Dependencies/DependenciesTests.swift
@@ -10,10 +10,7 @@ final class DependenciesTests: XCTestCase {
                 .github(path: "Dependency1/Dependency1", requirement: .branch("BranchName")),
                 .git(path: "Dependency2/Dependency2", requirement: .upToNext("1.2.3")),
             ],
-            swiftPackageManager: [
-                .local(path: "Path/Path"),
-                .remote(url: "Dependency3/Dependency3", requirement: .exact("4.5.6")),
-            ],
+            swiftPackageManager: .init(),
             platforms: [.iOS, .macOS, .tvOS, .watchOS]
         )
         XCTAssertCodable(subject)