+++ Tests/TuistCoreTests/Graph/GraphTraverserTests.swift
@@ -1310,7 +1310,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             )]),
             primaryBinaryPath: "/xcframeworks/c.xcframework/c",
             linking: .dynamic,
-            mergeable: false
+            mergeable: false,
+            status: .required
         )
         let dDependency = GraphDependency.xcframework(
             path: "/xcframeworks/d.xcframework",
@@ -1321,7 +1322,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             )]),
             primaryBinaryPath: "/xcframeworks/d.xcframework/d",
             linking: .dynamic,
-            mergeable: false
+            mergeable: false,
+            status: .required
         )
         let eDependency = GraphDependency.xcframework(
             path: "/xcframeworks/e.xcframework",
@@ -1333,7 +1335,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             )]),
             primaryBinaryPath: "/xcframeworks/e.xcframework/e",
             linking: .dynamic,
-            mergeable: true
+            mergeable: true,
+            status: .required
         )
         let dependencies: [GraphDependency: Set<GraphDependency>] = [
             .target(name: app.name, path: project.path): Set(arrayLiteral: cDependency, eDependency),
@@ -1379,7 +1382,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             )]),
             primaryBinaryPath: "/xcframeworks/c.xcframework/c",
             linking: .dynamic,
-            mergeable: false
+            mergeable: false,
+            status: .required
         )
         let dDependency = GraphDependency.xcframework(
             path: "/xcframeworks/d.xcframework",
@@ -1390,7 +1394,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             )]),
             primaryBinaryPath: "/xcframeworks/d.xcframework/d",
             linking: .dynamic,
-            mergeable: false
+            mergeable: false,
+            status: .required
         )
         let eDependency = GraphDependency.xcframework(
             path: "/xcframeworks/e.xcframework",
@@ -1402,7 +1407,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             )]),
             primaryBinaryPath: "/xcframeworks/e.xcframework/e",
             linking: .dynamic,
-            mergeable: true
+            mergeable: true,
+            status: .required
         )
         let dependencies: [GraphDependency: Set<GraphDependency>] = [
             .target(name: app.name, path: project.path): Set(arrayLiteral: cDependency, eDependency),
@@ -2688,7 +2694,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             bcsymbolmapPaths: [],
             linking: .dynamic,
             architectures: [.arm64],
-            isCarthage: false
+            isCarthage: false,
+            status: .required
         )
         let project = Project.test(path: "/path/project", targets: [app, staticFramework])
         let graph = Graph.test(
@@ -2724,7 +2731,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
                 bcsymbolmapPaths: [],
                 linking: .dynamic,
                 architectures: [.arm64],
-                product: .framework
+                product: .framework,
+                status: .required
             ),
         ])
     }
@@ -2741,7 +2749,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             bcsymbolmapPaths: [],
             linking: .dynamic,
             architectures: [.arm64],
-            isCarthage: false
+            isCarthage: false,
+            status: .required
         )
         let project = Project.test(path: "/path/project", targets: [app, framework])
         let graph = Graph.test(
@@ -2785,7 +2794,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             bcsymbolmapPaths: [],
             linking: .dynamic,
             architectures: [.arm64],
-            isCarthage: false
+            isCarthage: false,
+            status: .required
         )
         let project = Project.test(path: "/path/project", targets: [app, staticFramework, framework])
         let graph = Graph.test(
@@ -3977,7 +3987,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             )]),
             primaryBinaryPath: "/xcframeworks/direct.xcframework/direct",
             linking: .static,
-            mergeable: false
+            mergeable: false,
+            status: .required
         )
         let directFramework = GraphDependency.framework(
             path: "/frameworks/direct.framework",
@@ -3986,7 +3997,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             bcsymbolmapPaths: [],
             linking: .static,
             architectures: [.arm64],
-            isCarthage: false
+            isCarthage: false,
+            status: .required
         )
         let directFrameworkTarget = GraphDependency.target(name: staticFramework.name, path: project.path)
         let transitiveFrameworkTargetXCFramework = GraphDependency.xcframework(
@@ -3998,7 +4010,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             )]),
             primaryBinaryPath: "/xcframeworks/transitive-framework-target-xcframework.xcframework/transitive",
             linking: .static,
-            mergeable: false
+            mergeable: false,
+            status: .required
         )
         let transitiveXCFramework = GraphDependency.xcframework(
             path: "/xcframeworks/transitive.xcframework",
@@ -4009,7 +4022,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             )]),
             primaryBinaryPath: "/xcframeworks/transitive.xcframework/transitive",
             linking: .static,
-            mergeable: false
+            mergeable: false,
+            status: .required
         )
         let frameworkTransitiveXCFramework = GraphDependency.xcframework(
             path: "/xcframeworks/framework-transitive.xcframework",
@@ -4020,7 +4034,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
             )]),
             primaryBinaryPath: "/xcframeworks/framework-transitive.xcframework/framework-transitive",
             linking: .static,
-            mergeable: false
+            mergeable: false,
+            status: .required
         )
 
         let dependencies: [GraphDependency: Set<GraphDependency>] = [
@@ -4058,7 +4073,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
                     architectures: [.arm64]
                 )]),
                 primaryBinaryPath: "/xcframeworks/direct.xcframework/direct",
-                binaryPath: "/xcframeworks/direct.xcframework/direct"
+                binaryPath: "/xcframeworks/direct.xcframework/direct",
+                status: .required
             ),
             .xcframework(
                 path: "/xcframeworks/transitive.xcframework",
@@ -4068,7 +4084,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
                     architectures: [.arm64]
                 )]),
                 primaryBinaryPath: "/xcframeworks/transitive.xcframework/transitive",
-                binaryPath: "/xcframeworks/transitive.xcframework/transitive"
+                binaryPath: "/xcframeworks/transitive.xcframework/transitive",
+                status: .required
             ),
             .xcframework(
                 path: "/xcframeworks/framework-transitive.xcframework",
@@ -4078,7 +4095,8 @@ final class GraphTraverserTests: TuistUnitTestCase {
                     architectures: [.arm64]
                 )]),
                 primaryBinaryPath: "/xcframeworks/framework-transitive.xcframework/framework-transitive",
-                binaryPath: "/xcframeworks/framework-transitive.xcframework/framework-transitive"
+                binaryPath: "/xcframeworks/framework-transitive.xcframework/framework-transitive",
+                status: .required
             ),
         ].sorted())
     }