+++ Tests/TuistKitTests/Mappers/Factories/WorkspaceMapperFactoryTests.swift
@@ -104,19 +104,6 @@ final class WorkspaceMapperFactoryTests: TuistUnitTestCase {
         XCTAssertContainsElementOfType(got, LastUpgradeVersionWorkspaceMapper.self)
     }
 
-    func test_automation_contains_the_path_workspace_mapper() throws {
-        // Given
-        let workspaceDirectory = try temporaryPath()
-        subject = WorkspaceMapperFactory(projectMapper: SequentialProjectMapper(mappers: projectMapperFactory.default()))
-
-        // When
-        let got = subject.automation(workspaceDirectory: workspaceDirectory)
-
-        // Then
-        let mapper = XCTAssertContainsElementOfType(got, AutomationPathWorkspaceMapper.self)
-        XCTAssertEqual(mapper?.workspaceDirectory, workspaceDirectory)
-    }
-
     func test_cache_contains_the_generate_cacheable_schemes_workspace_mapper() throws {
         // Given
         let includedTargets = Set(arrayLiteral: "MyTarget")