+++ make/tasks/workspace/release/bundle.sh
@@ -7,6 +7,7 @@ ROOT_DIR=$($SCRIPT_DIR/../../../utilities/root_dir.sh)
 source $ROOT_DIR/make/utilities/setup.sh
 XCODE_PATH_SCRIPT_PATH=$SCRIPT_DIR/../../../utilities/xcode_path.sh
 BUILD_DIRECTORY=$ROOT_DIR/build
+export TUIST_INCLUDE_TUIST_CLOUD=1
 # Xcode 15 has a bug that causes the /var/folders... temporary directory, which is a symlink to
 # /private/var/folders to crash Xcode.
 TMP_DIR=/private$(mktemp -d)
@@ -17,7 +18,13 @@ BUILD_DIR=$ROOT_DIR/build
 
 echo "$(format_section "Building release into $BUILD_DIRECTORY")"
 
-XCODE_PATH=$($XCODE_PATH_SCRIPT_PATH --version $XCODE_VERSION)
+XCODE_PATH=/Applications/Xcode_15.0.app
+# TODO:
+# Remove the hardcoded path. The GitHub hosted runners have the dev tools of 15.0.1
+# and causes the line below to output a non-existing path. Once they've installed 15.0.1
+# in the CI environments, the hardcoded path should not be necessary.
+# That will happen when this PR is merged and deployed: https://github.com/actions/runner-images/pull/8601
+# XCODE_PATH=$($XCODE_PATH_SCRIPT_PATH --version $XCODE_VERSION)
 XCODE_LIBRARIES_PATH=$($XCODE_PATH_SCRIPT_PATH --version $LIBRARIES_XCODE_VERSION)
 
 echo "Static executables will be built with $XCODE_PATH"
@@ -67,20 +74,24 @@ build_fat_release_binary() {
         --configuration release \
         --disable-sandbox \
         --product $1 \
-        --build-path $TMP_DIR \
+        --package-path $2 \
+        --build-path $TMP_DIR/$1 \
         --triple $ARM64_TARGET
 
     DEVELOPER_DIR=$XCODE_PATH swift build \
         --configuration release \
         --disable-sandbox \
         --product $1 \
-        --build-path $TMP_DIR \
+        --package-path $2 \
+        --build-path $TMP_DIR/$1 \
         --triple $X86_64_TARGET
 
+    mkdir -p $3
+    
     DEVELOPER_DIR=$XCODE_PATH lipo -create \
-        -output $BUILD_DIRECTORY/$1 \
-        $TMP_DIR/$ARM64_TARGET/release/$1 \
-        $TMP_DIR/$X86_64_TARGET/release/$1
+        -output $3/$1 \
+        $TMP_DIR/$1/$ARM64_TARGET/release/$1 \
+        $TMP_DIR/$1/$X86_64_TARGET/release/$1
     )
 }
 
@@ -92,16 +103,16 @@ build_fat_release_library "ProjectDescription"
 echo "$(format_subsection "Building ProjectAutomation framework")"
 build_xcframework_library "ProjectAutomation"
 
-echo "$(format_subsection "Building tuist executable")"
-build_fat_release_binary "tuist"
+echo "$(format_subsection "Building xcbeautify executable")"
+build_fat_release_binary "xcbeautify" $ROOT_DIR/projects/tuist/vendor $BUILD_DIRECTORY/vendor
 
 echo "$(format_subsection "Building tuist executable")"
-build_fat_release_binary "tuistenv"
+build_fat_release_binary "tuist" $ROOT_DIR $BUILD_DIRECTORY
+
+echo "$(format_subsection "Building tuist executable")"
+build_fat_release_binary "tuistenv" $ROOT_DIR $BUILD_DIRECTORY
 
 echo "$(format_section "Copying assets")"
-
-echo "$(format_subsection "Copying Tuist's vendored resources")"
-cp -r $ROOT_DIR/projects/tuist/vendor $BUILD_DIRECTORY/vendor
 
 echo "$(format_subsection "Copying Tuist's templates")"
 cp -r $ROOT_DIR/Templates $BUILD_DIRECTORY/Templates