+++ projects/tuist/fixtures/app_with_plugins/Tuist/Package.swift
@@ -1,0 +1,7 @@
+// swift-tools-version: 5.8
+import PackageDescription
+
+let package = Package(
+    name: "PackageName",
+    dependencies: []
+)