+++ projects/tuist/fixtures/workspace_with_multiple_projects/Workspace.swift
@@ -1,0 +1,9 @@
+import ProjectDescription
+
+let workspace = Workspace(
+    name: "Tuist",
+    projects: [
+        "ProjectA",
+        "ProjectB",
+    ]
+)